requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
google-generativeai>=0.8.3
google-cloud-texttospeech>=2.31.0
pytest>=8.3.0
httpx>=0.27.0
